+++ src/pages/AdminPage.vue
@@ -10,6 +10,7 @@ import AdminConfig from '@/components/AdminConfig.vue'
 import AdminModulos from '@/components/AdminModulos.vue'
 import AdminFuncionarios from '@/components/AdminFuncionarios.vue'
 import AdminDisenos from '@/components/AdminDisenos.vue'
+import AdminDocumentos from '@/components/AdminDocumentos.vue'
 import AdminRubricaManager from '@/components/AdminRubricaManager.vue'
 import AdminCohortManager from '@/components/AdminCohortManager.vue'
 import AdminBadgeManager from '@/components/AdminBadgeManager.vue'
@@ -125,6 +126,7 @@ function handleDeleted(curso) {
 
       <AdminFuncionarios v-else-if="activeSection === 'funcionarios'" />
       <AdminDisenos v-else-if="activeSection === 'disenos'" />
+      <AdminDocumentos v-else-if="activeSection === 'documentos'" />
       <AdminModulos v-else-if="activeSection === 'modulos'" />
       <AdminConfig v-else-if="activeSection === 'config'" />